Extracted common `measureLabels` code to a shared component
