+++ packages/d3fc-axis/src/axisLabelOffset.js
@@ -1,26 +1,10 @@
 import { rebindAll, exclude } from '@d3fc/d3fc-rebind';
+import measureLabels from './measureLabels';
 
 export default (adaptee) => {
 
     let labelOffsetDepth = 'auto';
     let decorate = () => { };
-
-    const measureLabels = s => {
-        const scale = adaptee.scale();
-        const labels = scale['ticks'] ? scale.ticks() : scale.domain();
-
-        const tester = s.append('text');
-        const boundingBoxes = labels.map(l => tester.text(l).node().getBBox());
-        const maxHeight = Math.max(...boundingBoxes.map(b => b.height));
-        const maxWidth = Math.max(...boundingBoxes.map(b => b.width));
-        tester.remove();
-
-        return {
-            maxHeight,
-            maxWidth,
-            labelCount: labels.length
-        };
-    };
 
     const isVertical = () =>
         adaptee.orient() === 'left' || adaptee.orient() === 'right';
@@ -28,7 +12,7 @@ export default (adaptee) => {
     const sign = () => (adaptee.orient() === 'top' || adaptee.orient() === 'left') ? -1 : 1;
 
     const decorateOffset = sel => {
-        const { maxHeight, maxWidth, labelCount } = measureLabels(sel);
+        const { maxHeight, maxWidth, labelCount } = measureLabels(adaptee.scale())(sel);
         const range = adaptee.scale().range()[1];
 
         const offsetLevels = labelOffsetDepth === 'auto'
